requirements.txt
fastapi
uvicorn
playwright
mcp-server